+++ packages/bff/src/index.ts
@@ -1,2 +1,8 @@
-export { AppModule } from "./bootstrap/app.module";
+export { AppModule, createBffGatewayModule } from "./bootstrap/app.module";
 export { startBffServer } from "./bootstrap/main";
+export type {
+  MetaRegistryItem,
+  MetaRegistryProvider,
+  MetaResourceKind
+} from "./infra/integration/meta-registry.type";
+export type { RuntimeGatewayRunner } from "./controller/http/view.type";